pyproject.toml
[tool.poetry]
name = "second-brain"
version = "0.1.0"
description = "Provides capabilities for second brain Obsidian."
authors = ["pernulio <cpernul@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
pre-commit = "^3.6.0"
pyyaml = "^6.0.1"
markdown = "^3.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
